+++ packages/data/.mocharc.js
@@ -7,6 +7,9 @@ const { requireFrom } = require('@kitmi/sys');
 const pg = requireFrom('pg', process.cwd());
 runtime.loadModule('pg', pg);
 
+const rabbitmq = requireFrom('rabbitmq-client', process.cwd());
+runtime.loadModule('rabbitmq-client', rabbitmq);
+
 module.exports = {
     timeout: 300000,
     require: ['@kitmi/tester', './test/lib/helpers.js'],