+++ packages/data/src/drivers/index.js
@@ -1,1 +1,2 @@
 export * as postgres from './postgres';
+export * as rabbitmq from './rabbitmq';